requirements.txt
django
typesense
faker
psycopg2
argon2-cffi